basic login/loading page layout/functionality implemented
