+++ components/HeaderNav/HeaderNav.tsx
@@ -5,74 +5,11 @@ import {
   Heading,
   HStack,
   Button,
-  CircularProgress,
   Text,
+  Img,
 } from "@chakra-ui/react";
-import { useDisclosure } from "@chakra-ui/react";
 import { CloseIcon, HamburgerIcon } from "@chakra-ui/icons";
 import Link from "next/link";
-import { useRouter } from "next/router";
-
-import CeramicClient from "@ceramicnetwork/http-client";
-import ThreeIdResolver from "@ceramicnetwork/3id-did-resolver";
-import Web3 from "web3";
-import Web3Modal from "web3modal";
-import { TileDocument } from "@ceramicnetwork/stream-tile";
-
-import WalletConnectProvider from "@walletconnect/web3-provider";
-
-import { EthereumAuthProvider, ThreeIdConnect } from "@3id/connect";
-import { DID } from "dids";
-import { IDX } from "@ceramicstudio/idx";
-
-// network node that we're interacting with, can be local/prod
-// we're using a test network here
-const endpoint = "https://ceramic-clay.3boxlabs.com";
-
-export interface ProfileData {
-  content: {
-    identity: Identity;
-    accType: string;
-  };
-}
-
-export interface Identity {
-  firstName: string;
-  lastName: string;
-  email: string;
-  displayName: string;
-  bio: string;
-  twitter?: string;
-  linkedIn?: string;
-  website?: string;
-  businessName: string;
-  businessType: string;
-  businessLocation: string;
-  currTitle: string;
-  currLocation?: string;
-  functionalExpertise: any[];
-  industryExpertise: any[];
-  companyInfo?: CompanyInfo[];
-}
-
-export interface BasicProfile {
-  name: string;
-}
-export interface Profile {
-  identity: Identity;
-  name: string;
-  accType: string;
-}
-
-export interface CompanyInfo {
-  companyName: string;
-  companyTitle: string;
-  companyImg: any;
-  companyStart: Date;
-  companyEnd: Date;
-  companyFunc: string;
-  companyIndustry: string;
-}
 
 const HamburgerItem = ({ children, isLast, to = "/" }) => {
   return (
@@ -91,128 +28,49 @@ const HeaderNav = (props) => {
   const whichPage = props.whichPage;
   const [show, setShow] = useState(false);
   const toggleMenu = () => setShow(!show);
-  const [isSubmitted, setIsSubmitted] = React.useState(false);
-  const { isOpen, onOpen, onClose } = useDisclosure();
-  const [name, setName] = useState("");
-  const [email, setEmail] = useState("");
-  const [accType, setAccType] = useState("");
-  const [loaded, setLoaded] = useState(false);
-  const router = useRouter();
-
-  const handleWalletConnect = async () => {
-    const web3Modal = new Web3Modal({
-      disableInjectedProvider: false,
-      network: "rinkeby",
-      cacheProvider: false,
-      providerOptions: {
-        walletconnect: {
-          package: WalletConnectProvider,
-          options: {
-            rpc: {
-              1: "https://eth-rinkeby.alchemyapi.io/v2/QtLM8rW9nB6DobDu8KQx-7fYMS2rBlky",
-            },
-          },
-        },
-      },
-    });
-    const provider = await web3Modal.connect();
-    console.log(provider);
-    const web3 = new Web3(provider);
-    console.log(web3);
-    const accounts = await web3.eth.getAccounts();
-    const currAccount = accounts[0];
-    const ceramic = new CeramicClient(endpoint);
-    const idx = new IDX({ ceramic });
-    setIsSubmitted(true);
-    try {
-      const threeIdConnect = new ThreeIdConnect();
-      const ethProvider = new EthereumAuthProvider(
-        window.ethereum,
-        currAccount
-      );
-      await threeIdConnect.connect(ethProvider);
-
-      const did = new DID({
-        provider: threeIdConnect.getDidProvider(),
-        resolver: {
-          ...ThreeIdResolver.getResolver(ceramic),
-        },
-      });
-
-      ceramic.setDID(did);
-      await ceramic.did.authenticate();
-
-      // does not require signing to get user's public data
-      const data: BasicProfile = await idx.get(
-        "basicProfile",
-        `${currAccount}@eip155:1`
-      );
-      console.log("data: ", data);
-
-      const profileData: ProfileData = await TileDocument.deterministic(
-        ceramic,
-        { family: "user-profile-data" },
-        { anchor: false, publish: false }
-      );
-
-      console.log("profileData: ", profileData.content);
-      let identity = profileData.content.identity;
-      let profileAccType = profileData.content.accType;
-
-      if (data.name && identity.email && profileAccType) {
-        setName(data.name);
-        setEmail(identity.email);
-        setAccType(profileAccType);
-        setIsSubmitted(false);
-        router.push("/profile");
-      } else {
-        console.log("No profile, pls create one...");
-        router.push("/steps");
-      }
-    } catch (err) {
-      console.log("error: ", err);
-      router.push("/steps");
-      setLoaded(true);
-    }
-  };
 
   if (whichPage === "index") {
     return (
       <Flex
         h={10}
         mb={8}
-        p={10}
+        p={4}
         as="nav"
         align="center"
         justify="space-between"
         wrap="wrap"
         w="100%"
       >
-        <HStack spacing={10} alignItems="flex-start">
-          <Heading w="300px">Twali 👁‍🗨</Heading>
-        </HStack>
-        <Box
-          display={{ base: "block", md: "block" }}
-          flexBasis={{ base: "100%", md: "auto" }}
-        >
-          <Button
-            mb={{ base: 8, sm: 0 }}
-            mr={{ base: 0, sm: 8 }}
-            display="block"
-            onClick={handleWalletConnect}
-            colorScheme="teal"
+        <Img
+          width={"144px"}
+          height={"auto"}
+          src="/twali-assets/navbar_logo.png"
+        />
+        <HStack alignItems="center" w="130px" height={"32px"}>
+          <Flex
+            ml="2"
+            mt="1"
+            width={"100%"}
+            height={"100%"}
+            backgroundColor="transparent"
+            border={"1px solid #F9FFF2"}
+            alignItems={"center"}
+            justifyItems={"center"}
+            borderRadius={32}
           >
-            Connect Wallet{" "}
-            {isSubmitted ? (
-              <CircularProgress
-                size="22px"
-                thickness="4px"
-                isIndeterminate
-                color="#3C2E26"
-              />
-            ) : null}
-          </Button>
-        </Box>
+            <Text
+              color="white"
+              fontSize={"14px"}
+              margin={"auto"}
+              alignSelf={"center"}
+              fontWeight={"700"}
+              letterSpacing={"0.06em"}
+              textTransform={"uppercase"}
+            >
+              0xb794f...
+            </Text>
+          </Flex>
+        </HStack>
       </Flex>
     );
   } else if (whichPage === "sign-up" || whichPage === "steps") {
